test: Reorganize files
