+++ packages/aws-lambda-otel-extension/test/unit/event-detection.test.js
@@ -2,7 +2,7 @@
 
 const { expect } = require('chai');
 
-const { detectEventType } = require('../opt/otel-extension/eventDetection');
+const { detectEventType } = require('../../opt/otel-extension/eventDetection');
 
 describe('detectEventType', () => {
   it('should return alexa skill', () => {
