+++ packages/aws-lambda-otel-extension/test/utils/get-extension-server-mock.js
