+++ packages/aws-lambda-otel-extension/test/utils/normalize-otel-attributes.js
